+++ textx/scoping/rrel.py
@@ -45,7 +45,7 @@ def rrel_sequence():
 
 
 def rrel_expression():
-    return Optional("+m:"), rrel_sequence
+    return Optional(_(r'\+[mp]+:')), rrel_sequence
 
 
 def rrel_standalone():
@@ -56,7 +56,8 @@ class RRELBase(object):
     def __init__(self):
         pass
 
-    def get_next_matches(self, obj, lookup_list, allowed, first_element=False):
+    def get_next_matches(self, obj, lookup_list, allowed, matched_path,
+                         first_element=False):
         """
         This function yields potential matches encountered along the
         requested RREL.
@@ -83,15 +84,16 @@ class RRELBase(object):
         if not allowed(obj, lookup_list, self):  # also adjusts visited objs
             return  # recursion stopper
 
-        obj, lookup_list = self.apply(obj, lookup_list, first_element)
+        obj, lookup_list, matched_path = self.apply(
+            obj, lookup_list, matched_path, first_element)
         if obj is None:
             return
         elif isinstance(obj, list):
             for iobj in obj:
                 if iobj is not None:
-                    yield iobj, lookup_list
+                    yield iobj, lookup_list, matched_path
         else:
-            yield obj, lookup_list
+            yield obj, lookup_list, matched_path
 
 
 class RRELParent(RRELBase):
@@ -108,7 +110,7 @@ class RRELParent(RRELBase):
     def start_at_root(self):
         return False
 
-    def apply(self, obj, lookup_list, first_element):
+    def apply(self, obj, lookup_list, matched_path, first_element):
         """
         Args:
             obj: model object
@@ -122,8 +124,8 @@ class RRELParent(RRELBase):
         while hasattr(obj, "parent"):
             obj = obj.parent
             if textx_isinstance(obj, t):
-                return obj, lookup_list
-        return None, lookup_list
+                return obj, lookup_list, matched_path
+        return None, lookup_list, matched_path
 
 
 class RRELNavigation(RRELBase):
@@ -141,7 +143,7 @@ class RRELNavigation(RRELBase):
     def start_at_root(self):
         return True
 
-    def apply(self, obj, lookup_list, first_element):
+    def apply(self, obj, lookup_list, matched_path, first_element):
         """
         Args:
             obj: model object
@@ -157,33 +159,35 @@ class RRELNavigation(RRELBase):
             from textx import get_model
             obj = get_model(obj)
         if len(lookup_list) == 0 and self.consume_name:
-            return None, lookup_list
+            return None, lookup_list, matched_path
         if needs_to_be_resolved(obj, self.name):
-            return Postponed(), lookup_list
+            return Postponed(), lookup_list, matched_path
         if hasattr(obj, self.name):
             target = getattr(obj, self.name)
             if isinstance(target, list):
                 if not self.consume_name:
-                    return target, lookup_list  # return list
+                    return target, lookup_list, matched_path  # return list
                 else:
                     lst = list(filter(lambda x: hasattr(
                         x, "name") and getattr(
                         x, "name") == lookup_list[0], target))
                     if len(lst) > 0:
-                        return lst[0], lookup_list[1:]  # return obj
+                        return lst[0], lookup_list[1:], matched_path + [
+                            lst[0]]  # return obj
                     else:
-                        return None, lookup_list  # return None
+                        return None, lookup_list, matched_path  # return None
             else:
                 if not self.consume_name:
-                    return target, lookup_list
+                    return target, lookup_list, matched_path
                 else:
                     if hasattr(target, "name") and getattr(
                             target, "name") == lookup_list[0]:
-                        return target, lookup_list[1:]  # return obj
+                        return target, lookup_list[1:], matched_path + [
+                            target]  # return obj
                     else:
-                        return None, lookup_list  # return None
+                        return None, lookup_list, matched_path  # return None
         else:
-            return None, lookup_list
+            return None, lookup_list, matched_path
 
 
 class RRELBrackets(RRELBase):
@@ -201,12 +205,13 @@ class RRELBrackets(RRELBase):
     def __repr__(self):
         return '(' + str(self.seq) + ')'
 
-    def get_next_matches(self, obj, lookup_list, allowed, first_element=False):
+    def get_next_matches(self, obj, lookup_list, allowed, matched_path,
+                         first_element=False):
         if not allowed(obj, lookup_list, self):  # also adjusts visited objs
             return  # recursion stopper
-        for iobj, ilookup_list in self.seq.get_next_matches(
-                obj, lookup_list, allowed, first_element):
-            yield iobj, ilookup_list
+        for iobj, ilookup_list, imatched_path in self.seq.get_next_matches(
+                obj, lookup_list, allowed, matched_path, first_element):
+            yield iobj, ilookup_list, imatched_path
 
 
 class RRELDots(RRELBase):
@@ -223,7 +228,7 @@ class RRELDots(RRELBase):
     def start_at_root(self):
         return False
 
-    def apply(self, obj, lookup_list, first_element):
+    def apply(self, obj, lookup_list, matched_path, first_element):
         """
         Args:
             obj: model object
@@ -237,9 +242,9 @@ class RRELDots(RRELBase):
             obj = obj.parent
             num -= 1
         if num <= 1:
-            return obj, lookup_list
+            return obj, lookup_list, matched_path
         else:
-            return None, lookup_list
+            return None, lookup_list, matched_path
 
 
 class RRELSequence(RRELBase):
@@ -262,13 +267,14 @@ class RRELSequence(RRELBase):
             res = res or p.start_at_root()
         return res
 
-    def get_next_matches(self, obj, lookup_list, allowed, first_element=False):
+    def get_next_matches(self, obj, lookup_list, allowed, matched_path,
+                         first_element=False):
         if not allowed(obj, lookup_list, self):  # also adjusts visited objs
             return  # recursion stopper
         for ip in self.paths:
-            for iobj, ilookup_list in ip.get_next_matches(
-                    obj, lookup_list, allowed, first_element):
-                yield iobj, ilookup_list
+            for iobj, ilookup_list, imatched_path in ip.get_next_matches(
+                    obj, lookup_list, allowed, matched_path, first_element):
+                yield iobj, ilookup_list, imatched_path
 
 
 class RRELZeroOrMore(RRELBase):
@@ -289,43 +295,46 @@ class RRELZeroOrMore(RRELBase):
     def __repr__(self):
         return str(self.path_element) + '*'
 
-    def get_next_matches(self, obj, lookup_list, allowed, first_element=False):
+    def get_next_matches(self, obj, lookup_list, allowed, matched_path,
+                         first_element=False):
         assert isinstance(self.path_element, RRELBrackets)
         from textx.scoping import Postponed
 
-        def get_from_zero_or_more(obj, lookup_list, first_element=False):
+        def get_from_zero_or_more(obj, lookup_list, matched_path,
+                                  first_element=False):
             assert self.start_locally() or self.start_at_root()  # or, not xor
             if not allowed(obj, lookup_list, self):  # also adjusts visited objs
                 return  # recursion stopper
             if first_element:
                 if self.start_locally():
-                    yield obj, lookup_list
+                    yield obj, lookup_list, matched_path
                 if self.start_at_root():
                     from textx import get_model
-                    yield get_model(obj), lookup_list
+                    yield get_model(obj), lookup_list, matched_path
             else:
-                yield obj, lookup_list
+                yield obj, lookup_list, matched_path
             assert isinstance(self.path_element.seq, RRELSequence)
-            for iobj, ilookup_list in self.path_element.seq.get_next_matches(
-                    obj, lookup_list, allowed,
+            for iobj, ilookup_list,\
+                imatched_path in self.path_element.seq.get_next_matches(
+                    obj, lookup_list, allowed, matched_path,
                     first_element=first_element):
                 if isinstance(iobj, Postponed):
-                    yield iobj, ilookup_list  # found postponed
+                    yield iobj, ilookup_list, imatched_path  # found postponed
                     return
                 # yield from
-                for iiobj, iilookup_list in get_from_zero_or_more(
-                        iobj, ilookup_list):
-                    yield iiobj, iilookup_list
+                for iiobj, iilookup_list, iimatched_path in get_from_zero_or_more(
+                        iobj, ilookup_list, imatched_path):
+                    yield iiobj, iilookup_list, iimatched_path
 
         prevent_doubles = set()
-        for iobj, ilookup_list in get_from_zero_or_more(
-                obj, lookup_list, first_element):
+        for iobj, ilookup_list, imatched_path in get_from_zero_or_more(
+                obj, lookup_list, matched_path, first_element):
             if isinstance(iobj, Postponed):
-                yield iobj, ilookup_list
+                yield iobj, ilookup_list, imatched_path
                 return
             if (id(iobj), len(ilookup_list)) not in prevent_doubles:
                 prevent_doubles.add((id(iobj), len(ilookup_list)))
-                yield iobj, ilookup_list
+                yield iobj, ilookup_list, imatched_path
 
 
 class RRELPath(RRELBase):
@@ -350,38 +359,45 @@ class RRELPath(RRELBase):
     def start_at_root(self):
         return self.path_elements[0].start_at_root()
 
-    def get_next_matches(self, obj, lookup_list, allowed, first_element=False):
+    def get_next_matches(self, obj, lookup_list, allowed, matched_path,
+                         first_element=False):
         from textx.scoping import Postponed
 
-        def intern_get_next_matches(obj, lookup_list, allowed, first_element=False,
-                                    idx=0):
+        def intern_get_next_matches(obj, lookup_list, allowed, matched_path,
+                                    first_element=False, idx=0):
             assert len(self.path_elements) > idx
             e = self.path_elements[idx]
-            for iobj, ilookup_list in e.get_next_matches(
-                    obj, lookup_list, allowed, first_element):
+            for iobj, ilookup_list, imatched_path in e.get_next_matches(
+                    obj, lookup_list, allowed, matched_path, first_element):
                 if isinstance(iobj, Postponed):
-                    yield iobj, ilookup_list
+                    yield iobj, ilookup_list, imatched_path
                     return
                 if len(self.path_elements) - 1 == idx:
-                    yield iobj, ilookup_list
+                    yield iobj, ilookup_list, imatched_path
                 else:
-                    for iiobj, iilookup_list in intern_get_next_matches(
-                            iobj, ilookup_list, allowed,
+                    for iiobj, iilookup_list,\
+                        iimatched_path in intern_get_next_matches(
+                            iobj, ilookup_list, allowed, imatched_path,
                             first_element=False, idx=idx + 1):
-                        yield iiobj, iilookup_list
+                        yield iiobj, iilookup_list, iimatched_path
 
-        for iobj, ilookup_list in intern_get_next_matches(
-                obj, lookup_list, allowed, first_element, 0):
-            yield iobj, ilookup_list
+        for iobj, ilookup_list, imatched_path in intern_get_next_matches(
+                obj, lookup_list, allowed, matched_path, first_element, 0):
+            yield iobj, ilookup_list, imatched_path
 
 
 class RRELExpression:
-    def __init__(self, seq, importURI):
+    def __init__(self, seq, flags):
         self.seq = seq
-        self.importURI = importURI
+        self.flags = flags
+        self.importURI = ('m' in flags)
+        self.use_proxy = ('p' in flags)
 
     def __repr__(self):
-        return "+m:" + str(self.seq) if self.importURI else str(self.seq)
+        if self.importURI:
+            return "+" + self.flags + ":" + str(self.seq)
+        else:
+            return str(self.seq)
 
 
 class RRELVisitor(PTNodeVisitor):
@@ -417,9 +433,53 @@ class RRELVisitor(PTNodeVisitor):
 
     def visit_rrel_expression(self, node, children):
         if len(children) == 1:
-            return RRELExpression(children[0], False)
+            return RRELExpression(children[0], "")
         else:
-            return RRELExpression(children[1], True)
+            flags = children[0][1:-1]  # see grammar
+            return RRELExpression(children[1], flags)
+
+
+class ReferenceProxy(object):
+    """
+    This class describes a resolved reference of an RREL
+    with 'p'-flag set ('+p:'). It can be used similar to
+    a normally resolved reference concerning the following
+    semantics: getattr, setattr, delattr and '=='.
+
+    With this, a ReferenceProxy object is compatible with the
+    `textx.textx_isinstance` semantics.
+
+    It further provides
+     - an attribute '_tx_obj', corresponding
+       exactly to the resolved reference.
+     - an attribute '_tx_path', a list of objects
+       describing the path of named objects traversed
+       during reference resolution.
+    """
+    def __init__(self, path):
+        assert len(path) > 0
+        self.__dict__['_tx_path'] = path
+
+    def __setattr__(self, key, value):
+        if key == '_tx_path' or key == '_tx_obj':
+            raise Exception("not allowed to set _tx_path or _tx_obj")
+        return setattr(self.__dict__['_tx_path'][-1], key, value)
+
+    def __getattr__(self, item):
+        if item == '_tx_path':
+            return self.__dict__['_tx_path']
+        elif item == '_tx_obj':
+            return self.__dict__['_tx_path'][-1]
+        else:
+            return getattr(self.__dict__['_tx_path'][-1], item)
+
+    def __delattr__(self, item):
+        if item == '_tx_path' or item == '_tx_obj':
+            raise Exception("not allowed to del _tx_path or _tx_obj")
+        return delattr(self.__dict__['_tx_path'][-1], item)
+
+    def __eq__(self, other):
+        return self.__dict__['_tx_path'][-1] == other
 
 
 def parse(rrel_expression):
@@ -438,23 +498,8 @@ def parse(rrel_expression):
     return visit_parse_tree(parse_tree, RRELVisitor())
 
 
-def find(obj, lookup_list, rrel_tree, obj_cls=None, split_string="."):
-    """
-    This function gets all/one element from a model
-    object based on an rrel tree (query).
-
-    Args:
-        obj: model object (starting point of the search)
-        lookup_list: list of name parts forming the full name.
-        rrel_tree: the query (must be a RRELExpression object or a string)
-        split_string: the string used to split the name into individual
-            parts (e.g. '.' for a python-like name schema or '::' for a
-            C++-like name schema for namespace resolution)
-
-    Returns:
-        The result of the query (first match), a
-        Postponed object, or None (nothing found)
-    """
+def find_object_with_path(obj, lookup_list, rrel_tree, obj_cls=None,
+                          split_string="."):
     from textx import textx_isinstance
     from textx.scoping import Postponed
 
@@ -476,14 +521,45 @@ def find(obj, lookup_list, rrel_tree, obj_cls=None, split_string="."):
             return True
 
     for p in rrel_tree.paths:
-        for obj_res, lookup_list_res in p.get_next_matches(
-                obj, lookup_list, allowed, first_element=True):
+        for obj_res, lookup_list_res, matched_path in p.get_next_matches(
+                obj, lookup_list, allowed, [], first_element=True):
             if isinstance(obj_res, Postponed):
                 return obj_res  # Postponed
             elif len(lookup_list_res) == 0:
                 if obj_cls is None or textx_isinstance(obj_res, obj_cls):
-                    return obj_res  # found match
+                    return obj_res, matched_path  # found match
     return None  # not found
+
+
+def find(obj, lookup_list, rrel_tree, obj_cls=None, split_string=".",
+         use_proxy=False):
+    """
+    This function gets all/one element from a model
+    object based on an rrel tree (query).
+
+    Args:
+        obj: model object (starting point of the search)
+        lookup_list: list of name parts forming the full name.
+        rrel_tree: the query (must be a RRELExpression object or a string)
+        split_string: the string used to split the name into individual
+            parts (e.g. '.' for a python-like name schema or '::' for a
+            C++-like name schema for namespace resolution)
+        use_proxy: return a path proxy object (see ReferenceProxy) instead
+            of the normal result of the query.
+
+    Returns:
+        The result of the query (first match), a
+        Postponed object, or None (nothing found)
+    """
+    res = find_object_with_path(obj, lookup_list, rrel_tree, obj_cls, split_string)
+    if type(res) is tuple:
+        # full path is in res[1]
+        if use_proxy:
+            return ReferenceProxy(res[1])
+        else:
+            return res[0]
+    else:
+        return res
 
 
 def create_rrel_scope_provider(rrel_tree_or_string, split_string=None, **kwargs):
@@ -507,7 +583,7 @@ def create_rrel_scope_provider(rrel_tree_or_string, split_string=None, **kwargs)
         """
         RREL scope provider
         """
-        def __init__(self, rrel_tree, split_string):
+        def __init__(self, rrel_tree, split_string, use_proxy):
             """
             Creates a RREL scope provider
 
@@ -515,9 +591,11 @@ def create_rrel_scope_provider(rrel_tree_or_string, split_string=None, **kwargs)
                 rrel_tree: the query (see `rrel.find`)
                 split_string: the string used to split the name into individual
                     parts (see `create_rrel_scope_provider`)
+                use_proxy: see find
             """
             self.rrel_tree = rrel_tree
             self.split_string = split_string
+            self.use_proxy = use_proxy
 
         def __call__(self, current_obj, attr, obj_ref):
             """
@@ -544,17 +622,17 @@ def create_rrel_scope_provider(rrel_tree_or_string, split_string=None, **kwargs)
                 split = self.split_string
 
             return find(current_obj, obj_name, self.rrel_tree, obj_cls,
-                        split_string=split)
+                        split_string=split, use_proxy=self.use_proxy)
 
     class RRELImportURI(ImportURI):
         """
         scope provider with ImportURI and RREL
         """
 
-        def __init__(self, rrel_tree, split_string,
+        def __init__(self, rrel_tree, split_string, use_proxy,
                      glob_args=None, search_path=None, importAs=False,
                      importURI_converter=None, importURI_to_scope_name=None):
-            ImportURI.__init__(self, RREL(rrel_tree, split_string),
+            ImportURI.__init__(self, RREL(rrel_tree, split_string, use_proxy),
                                glob_args=glob_args,
                                search_path=search_path, importAs=importAs,
                                importURI_converter=importURI_converter,
@@ -564,6 +642,8 @@ def create_rrel_scope_provider(rrel_tree_or_string, split_string=None, **kwargs)
         rrel_tree_or_string = parse(rrel_tree_or_string)
 
     if rrel_tree_or_string.importURI:
-        return RRELImportURI(rrel_tree_or_string, split_string, *kwargs)
+        return RRELImportURI(rrel_tree_or_string, split_string,
+                             rrel_tree_or_string.use_proxy, *kwargs)
     else:
-        return RREL(rrel_tree_or_string, split_string)
+        return RREL(rrel_tree_or_string, split_string,
+                    rrel_tree_or_string.use_proxy)